Fix issue with tiptap having duplicate plugin keys
- Update tiptap and its extensions to latest version, 2.4.0

- Pin vuex version to 3.6.x
  This avoids npm installing the latest vuex version, vuex4. Current
installed `vuex-persist` package is not compatible with vuex4. As of
this writing, 3.6.3 is the latest vuex3 version.

- Regenerated package-lock.json
  Weird JS dependency issue called for removal of `node_modules` and
regeneration of package-lock.json. Resulting on some packages to be
upgraded.
This might need more testing to make sure other parts of the website is
not affected by more js dependency compatibility issues.